:recycle: refactor(chat): 도메인 기준으로 디렉토리 구조 변경
feat/chat
