+++ src/main/java/com/jupjup/www/jupjup/chat/repository/RoomRepository.java
@@ -1,10 +1,8 @@
-package com.jupjup.www.jupjup.domain.repository.chat;
+package com.jupjup.www.jupjup.chat.repository;
 
-import com.jupjup.www.jupjup.domain.entity.chat.Room;
+import com.jupjup.www.jupjup.chat.entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface RoomRepository extends JpaRepository<Room, Long> {
