+++ src/main/java/com/jupjup/www/jupjup/chat/repository/UserChatRoomRepository.java
@@ -1,6 +1,6 @@
-package com.jupjup.www.jupjup.domain.repository.chat;
+package com.jupjup.www.jupjup.chat.repository;
 
-import com.jupjup.www.jupjup.domain.entity.chat.UserChatRoom;
+import com.jupjup.www.jupjup.chat.entity.UserChatRoom;
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
