+++ src/main/java/com/jupjup/www/jupjup/domain/entity/giveaway/Giveaway.java
@@ -1,7 +1,7 @@
 package com.jupjup.www.jupjup.domain.entity.giveaway;
 
 import com.jupjup.www.jupjup.domain.entity.User;
-import com.jupjup.www.jupjup.domain.entity.chat.Room;
+import com.jupjup.www.jupjup.chat.entity.Room;
 import com.jupjup.www.jupjup.domain.enums.GiveawayStatus;
 import com.jupjup.www.jupjup.image.entity.Image;
 import jakarta.persistence.*;
